refactor trocado lib noSql para jpa feat incluindo kotlin
