+++ src/main/java/br/com/m3Tech/solucoesFromtis/controller/CnabRetornoCobrancaController.java
@@ -47,299 +47,295 @@ import lombok.Setter;
 @Controller
 public class CnabRetornoCobrancaController implements Serializable {
 
-	private static final long serialVersionUID = 1L;
-	
-
-	private static final String VOLTAR = "/pages/cadastros/index.xhtml";
+    private static final long serialVersionUID = 1L;
 
 
-	@Autowired
-	private IBaseService baseService;
-	@Autowired
-	private IFundoService fundoService;
-	@Autowired
-	private IBancoService bancoService;
-	@Autowired
-	private IMovimentoService movimentoService;
-	@Autowired
-	private IConfGlobalService confGlobalService;
-	@Autowired
-	private  IGeradorCnab geradorCnab;
-		
-	
-	
-	private Integer baseSelecionada;
-	private Integer fundoSelecionado;
-	private Integer bancoSelecionado;
-	private Integer layoutSelecionado;
-	private Integer movimentoSelecionado;
-	
-	private String path;
-	private String agenciaCobranca = StringUtils.getNumeroComZerosAEsquerda(ValorAleatorioUtil.getValorNumerico(9999), 4);
-	private String idCobranca = StringUtils.getNumeroComZerosAEsquerda(ValorAleatorioUtil.getValorNumerico(999999), 6);
-	
-	private Boolean importacaoAutomatica;
-	
-	private LocalDate dataGravacao;
-	
-	private BigDecimal valorPago;
-	private BigDecimal valorAbatimento = BigDecimal.ZERO;
-	private BigDecimal valorDesconto = BigDecimal.ZERO;
-	private BigDecimal valorJuros = BigDecimal.ZERO;
-	private BigDecimal valorDespesas = BigDecimal.ZERO;
-	
-	private CnabCobrancaDto cnab;
-	
-	private List<Base> bases;
-	private List<FundoDto> fundos;
-	private List<BancoDto> bancos;
-	private List<OriginadorDto> originadores;
-	private List<MovimentoDto> movimentos;
-	private List<TituloDto> titulos;
-	private List<TituloDto> titulosEmEstoque;
-	private List<LayoutEnum> layoutsCobranca;
-	private List<FundoCobrancaDto> cobrancas;
-	
-	@PostConstruct
-	public void init() {
-		
-		bases = baseService.findAll();
-		fundos = new ArrayList<>();
-		bancos = new ArrayList<>();
-		originadores = new ArrayList<>();
-		movimentos = new ArrayList<>();
-		titulosEmEstoque = new ArrayList<>();
-		cobrancas = new ArrayList<>();
-		cnab = new CnabCobrancaDto();
-		path = confGlobalService.getConfGlobal().getPath();
+    private static final String VOLTAR = "/pages/cadastros/index.xhtml";
 
-				
-	}
-	
-	public void selecionandoBase() {
-		System.out.println("Selecionando Base " + baseSelecionada );
-		cnab = new CnabCobrancaDto();
-		atualizarFundos();
-		selecionandoLayout();
-		
-	}
 
-	
-	public void selecionandoFundo() {
-		System.out.println("Selecionando Fundo " + fundoSelecionado );
-		cnab = new CnabCobrancaDto();
-		Optional<FundoDto> optional = fundos.stream().filter(f -> f.getIdFundo().equals(fundoSelecionado)).findFirst();
-		
-		if(optional.isPresent()) {
-			try {
-				Base base = baseService.findById(baseSelecionada);
-				
-				Connection con = Conexao.getConnection(base);
-				FundoDto fundo = optional.get();
+    @Autowired
+    private IBaseService baseService;
+    @Autowired
+    private IFundoService fundoService;
+    @Autowired
+    private IBancoService bancoService;
+    @Autowired
+    private IMovimentoService movimentoService;
+    @Autowired
+    private IConfGlobalService confGlobalService;
+    @Autowired
+    private IGeradorCnab geradorCnab;
 
-				dataGravacao = fundo.getDataFundo();				
-				selecionandoLayout(); 
+
+    private Integer baseSelecionada;
+    private Integer fundoSelecionado;
+    private Integer bancoSelecionado;
+    private Integer layoutSelecionado;
+    private Integer movimentoSelecionado;
+
+    private String path;
+    private String agenciaCobranca = StringUtils.getNumeroComZerosAEsquerda(ValorAleatorioUtil.getValorNumerico(9999), 4);
+    private String idCobranca = StringUtils.getNumeroComZerosAEsquerda(ValorAleatorioUtil.getValorNumerico(999999), 6);
+
+    private Boolean importacaoAutomatica;
+
+    private LocalDate dataGravacao;
+
+    private BigDecimal valorPago;
+    private BigDecimal valorAbatimento = BigDecimal.ZERO;
+    private BigDecimal valorDesconto = BigDecimal.ZERO;
+    private BigDecimal valorJuros = BigDecimal.ZERO;
+    private BigDecimal valorDespesas = BigDecimal.ZERO;
+
+    private CnabCobrancaDto cnab;
+
+    private List<Base> bases;
+    private List<FundoDto> fundos;
+    private List<BancoDto> bancos;
+    private List<OriginadorDto> originadores;
+    private List<MovimentoDto> movimentos;
+    private List<TituloDto> titulos;
+    private List<TituloDto> titulosEmEstoque;
+    private List<LayoutEnum> layoutsCobranca;
+    private List<FundoCobrancaDto> cobrancas;
+
+    @PostConstruct
+    public void init() {
+
+        bases = baseService.findAll();
+        fundos = new ArrayList<>();
+        bancos = new ArrayList<>();
+        originadores = new ArrayList<>();
+        movimentos = new ArrayList<>();
+        titulosEmEstoque = new ArrayList<>();
+        cobrancas = new ArrayList<>();
+        cnab = new CnabCobrancaDto();
+        path = confGlobalService.getConfGlobal().getPath();
+
+
+    }
+
+    public void selecionandoBase() {
+        System.out.println("Selecionando Base " + baseSelecionada);
+        cnab = new CnabCobrancaDto();
+        atualizarFundos();
+        selecionandoLayout();
+
+    }
+
+
+    public void selecionandoFundo() {
+        System.out.println("Selecionando Fundo " + fundoSelecionado);
+        cnab = new CnabCobrancaDto();
+        Optional<FundoDto> optional = fundos.stream().filter(f -> f.getIdFundo().equals(fundoSelecionado)).findFirst();
+
+        if (optional.isPresent()) {
+            try {
+                Base base = baseService.findById(baseSelecionada);
+
+                Connection con = Conexao.getConnection(base);
+                FundoDto fundo = optional.get();
+
+                dataGravacao = fundo.getDataFundo();
+                selecionandoLayout();
 //				movimentos = movimentoService.findAllMovimentos(con, layoutSelecionado);
 //				atualizarTitulosEmEstoque();
-				
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-				
-	}
 
-	
-	
-	public void selecionandoLayout() {
-		try {
-			Base base = baseService.findById(baseSelecionada);
-			
-			layoutsCobranca = new ArrayList<>();
-			bancos = new ArrayList<>();
-			movimentos = new ArrayList<>();
-			
-			if(IntegerUtils.isZeroOrNull(fundoSelecionado)) {
-				return;
-			}
-			
-			cobrancas = fundoService.findCodCobrancas( Conexao.getConnection(base),fundoSelecionado);
-			
-			cobrancas.forEach(c -> {
-				try {
-					LayoutEnum layout = LayoutEnum.findCobrancaByCodBanco(c.getNuBanco());
-					if(!layoutsCobranca.contains(layout)) {
-						layoutsCobranca.add(layout);
-					}
-					
-					BancoDto banco = bancoService.findOneByNumBanco(Conexao.getConnection(base), c.getNuBanco());
-					if(!bancos.contains(banco)) {
-						bancos.add(banco);
-					}
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-				
-				});
-			
-			movimentos = movimentoService.findAllMovimentos(Conexao.getConnection(base), layoutSelecionado == null ? layoutsCobranca.get(0).getCdLayout():layoutSelecionado);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+
+    }
+
+
+    public void selecionandoLayout() {
+        try {
+            Base base = baseService.findById(baseSelecionada);
+
+            layoutsCobranca = new ArrayList<>();
+            bancos = new ArrayList<>();
+            movimentos = new ArrayList<>();
+
+            if (IntegerUtils.isZeroOrNull(fundoSelecionado)) {
+                return;
+            }
+
+            cobrancas = fundoService.findCodCobrancas(Conexao.getConnection(base), fundoSelecionado);
+
+            cobrancas.forEach(c -> {
+                try {
+                    LayoutEnum layout = LayoutEnum.findCobrancaByCodBanco(c.getNuBanco());
+                    if (!layoutsCobranca.contains(layout)) {
+                        layoutsCobranca.add(layout);
+                    }
+
+                    BancoDto banco = bancoService.findOneByNumBanco(Conexao.getConnection(base), c.getNuBanco());
+                    if (!bancos.contains(banco)) {
+                        bancos.add(banco);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+
+            });
+
+            movimentos = movimentoService.findAllMovimentos(Conexao.getConnection(base), layoutSelecionado == null ? layoutsCobranca.get(0).getCdLayout() : layoutSelecionado);
 //			atualizarTitulosEmEstoque(); 
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	public void selecionandoImportacaoAutomatica() {
-		
-		try {
-			Preconditions.checkNotNull(baseSelecionada, "É obrigatório selecionar uma base");
-			Preconditions.checkNotNull(fundoSelecionado, "É obrigatório selecionar uma fundo");
-			
-			Base base = baseService.findById(baseSelecionada);
-			Optional<FundoDto> optional = fundos.stream().filter(c -> c.getIdFundo().equals(fundoSelecionado)).findFirst();
-			
-			if(optional.isPresent()) {
-				FundoDto fundoSelecionado = optional.get();
-				
-				path = (confGlobalService.getPathSalvarArquivo(Conexao.getConnection(base), importacaoAutomatica, base.getVersaoMercado(), fundoSelecionado));
-			}
-			
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
 
-		} catch (Exception e1) {
-			e1.printStackTrace();
-		}
-		
-	}
-	
-	private void atualizarFundos() {
-		
-		try {
-		
-			Base base = baseService.findById(baseSelecionada);
-		
-			fundos = fundoService.findAll(Conexao.getConnection(base));
-			
-			selecionandoLayout();
-			
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
+    public void selecionandoImportacaoAutomatica() {
 
-		
-	}
-	
-	
-	public void atualizarTitulosEmEstoque() {
+        try {
+            Preconditions.checkNotNull(baseSelecionada, "É obrigatório selecionar uma base");
+            Preconditions.checkNotNull(fundoSelecionado, "É obrigatório selecionar uma fundo");
 
-		try {
-			Base base = baseService.findById(baseSelecionada);
-			
-			if(movimentoSelecionado == null) {
-				return;
-			}
-			
-			Optional<MovimentoDto> optionalMovimento = movimentos.stream().filter(c -> c.getIdMovimento().equals(movimentoSelecionado)).findFirst();
-			
-			if(optionalMovimento.isPresent()) {
-				
-				MovimentoDto movimentoDto = optionalMovimento.get();
-			
-				titulosEmEstoque = movimentoService.findAllTituloEmEstoqueCobranca( Conexao.getConnection(base),fundoSelecionado,movimentoDto.getIdMovimento());
-			
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	
-	public void atualizarBases() {
-		System.out.println("Atualizar Base");
-		
-		bases = baseService.findAll();
-	}
-	
-	public void addTitulo(TituloDto item) {
-		
-		Optional<MovimentoDto> optionalMovimento = movimentos.stream().filter(c -> c.getIdMovimento().equals(movimentoSelecionado)).findFirst();
+            Base base = baseService.findById(baseSelecionada);
+            Optional<FundoDto> optional = fundos.stream().filter(c -> c.getIdFundo().equals(fundoSelecionado)).findFirst();
 
-		if(optionalMovimento.isPresent()) {
-			item.setMovimento(optionalMovimento.get());
-		}
-		
-		item.setAgenciaCobranca(agenciaCobranca);
-		item.setIdCobranca(idCobranca);
-		
-		
-		
-		item.setValorPago(valorPago); 
-		item.setValorAbatimento(valorAbatimento);
-		item.setValorDesconto(valorDesconto);
-		item.setValorJuros(valorJuros);
-		item.setValorDespesas(valorDespesas);
-		
-		cnab.getTitulos().add(item.getCopy());
-		titulosEmEstoque.remove(item);
-		
-		clear();
-		
-	}
-	
-	public void gerar() {
-		try {
-			
-			if(cnab.getTitulos() == null || cnab.getTitulos().isEmpty()) {
-				FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_ERROR, "", "Nenhum título foi adicionado.."));
-				return;
-			}
-			
-			BancoDto banco = bancos.stream().filter(c -> c.getIdBanco().equals(bancoSelecionado)).findFirst().get();
-			
-			cnab.setBanco(banco);
-			cnab.setDataGravacao(dataGravacao);
-			cnab.setFundo(fundos.stream().filter(c -> c.getIdFundo().equals(fundoSelecionado)).findFirst().get());
-			cnab.setLayout(LayoutEnum.parse(layoutSelecionado));
-			
-			
-			FundoCobrancaDto fundoCobrancaDto = cobrancas.stream().filter(c -> c.getNuBanco().equalsIgnoreCase(banco.getCodigoBanco())).findFirst().get();
-			
-			cnab.setCodigoCobranca(fundoCobrancaDto.getCodCobranca());
-			
-			
-			ConfGlobal confGlobal = confGlobalService.getConfGlobal();
-			cnab.setNumSeqArquivo(confGlobal.getSeqArquivo());
-			confGlobal.setSeqArquivo(confGlobal.getSeqArquivo() + 1);
-			if(!importacaoAutomatica) {
-				confGlobal.setPath(path);
-			}
-		
-		
-//			confGlobal.save();TODO
-			
-			geradorCnab.gerarRetornoCobrança(cnab, "RETORNO_COBRANCA", importacaoAutomatica, path);
-			
-			FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "", "Cnab Gerado Com Sucesso."));
-			
-			cnab = new CnabCobrancaDto();
-			
-			clear();
-			
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	private void clear() {
+            if (optional.isPresent()) {
+                FundoDto fundoSelecionado = optional.get();
 
-	}
+                path = (confGlobalService.getPathSalvarArquivo(Conexao.getConnection(base), importacaoAutomatica, base.getVersaoMercado(), fundoSelecionado));
+            }
 
-	
-	public void removerTitulo(TituloDto titulo) {
-		cnab.getTitulos().remove(titulo);
-	}
 
-	public String voltar() {
-		return VOLTAR;
-	}
+        } catch (Exception e1) {
+            e1.printStackTrace();
+        }
+
+    }
+
+    private void atualizarFundos() {
+
+        try {
+
+            Base base = baseService.findById(baseSelecionada);
+
+            fundos = fundoService.findAll(Conexao.getConnection(base));
+
+            selecionandoLayout();
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+
+    }
+
+
+    public void atualizarTitulosEmEstoque() {
+
+        try {
+            Base base = baseService.findById(baseSelecionada);
+
+            if (movimentoSelecionado == null) {
+                return;
+            }
+
+            Optional<MovimentoDto> optionalMovimento = movimentos.stream().filter(c -> c.getIdMovimento().equals(movimentoSelecionado)).findFirst();
+
+            if (optionalMovimento.isPresent()) {
+
+                MovimentoDto movimentoDto = optionalMovimento.get();
+
+                titulosEmEstoque = movimentoService.findAllTituloEmEstoqueCobranca(Conexao.getConnection(base), fundoSelecionado, movimentoDto.getIdMovimento());
+
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+
+    public void atualizarBases() {
+        System.out.println("Atualizar Base");
+
+        bases = baseService.findAll();
+    }
+
+    public void addTitulo(TituloDto item) {
+
+        Optional<MovimentoDto> optionalMovimento = movimentos.stream().filter(c -> c.getIdMovimento().equals(movimentoSelecionado)).findFirst();
+
+        if (optionalMovimento.isPresent()) {
+            item.setMovimento(optionalMovimento.get());
+        }
+
+        item.setAgenciaCobranca(agenciaCobranca);
+        item.setIdCobranca(idCobranca);
+
+
+        item.setValorPago(valorPago);
+        item.setValorAbatimento(valorAbatimento);
+        item.setValorDesconto(valorDesconto);
+        item.setValorJuros(valorJuros);
+        item.setValorDespesas(valorDespesas);
+
+        cnab.getTitulos().add(item.getCopy());
+        titulosEmEstoque.remove(item);
+
+        clear();
+
+    }
+
+    public void gerar() {
+        try {
+
+            if (cnab.getTitulos() == null || cnab.getTitulos().isEmpty()) {
+                FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_ERROR, "", "Nenhum título foi adicionado.."));
+                return;
+            }
+
+            BancoDto banco = bancos.stream().filter(c -> c.getIdBanco().equals(bancoSelecionado)).findFirst().get();
+
+            cnab.setBanco(banco);
+            cnab.setDataGravacao(dataGravacao);
+            cnab.setFundo(fundos.stream().filter(c -> c.getIdFundo().equals(fundoSelecionado)).findFirst().get());
+            cnab.setLayout(LayoutEnum.parse(layoutSelecionado));
+
+
+            FundoCobrancaDto fundoCobrancaDto = cobrancas.stream().filter(c -> c.getNuBanco().equalsIgnoreCase(banco.getCodigoBanco())).findFirst().get();
+
+            cnab.setCodigoCobranca(fundoCobrancaDto.getCodCobranca());
+
+
+            ConfGlobal confGlobal = confGlobalService.getConfGlobal();
+            cnab.setNumSeqArquivo(confGlobal.getSeqArquivo());
+            confGlobal.setSeqArquivo(confGlobal.getSeqArquivo() + 1);
+            if (!importacaoAutomatica) {
+                confGlobal.setPath(path);
+            }
+
+            this.confGlobalService.salvar(confGlobal);
+
+            geradorCnab.gerarRetornoCobrança(cnab, "RETORNO_COBRANCA", importacaoAutomatica, path);
+
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "", "Cnab Gerado Com Sucesso."));
+
+            cnab = new CnabCobrancaDto();
+
+            clear();
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void clear() {
+
+    }
+
+
+    public void removerTitulo(TituloDto titulo) {
+        cnab.getTitulos().remove(titulo);
+    }
+
+    public String voltar() {
+        return VOLTAR;
+    }
 
 }
