+++ src/main/java/br/com/m3Tech/solucoesFromtis/model/Base.java
@@ -6,10 +6,13 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.jpa.domain.AbstractPersistable;
 
+import javax.persistence.Entity;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Entity
 public class Base extends AbstractPersistable<Integer> {
 
 	private String url;
