+++ src/main/java/br/com/m3Tech/solucoesFromtis/repositories/BaseRepository.kt
@@ -1,0 +1,7 @@
+package br.com.m3Tech.solucoesFromtis.repositories
+
+import br.com.m3Tech.solucoesFromtis.model.Base
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface BaseRepository : JpaRepository<Base, Integer> {
+}
