+++ src/main/java/br/com/m3Tech/solucoesFromtis/repositories/ConfGlobalRepository.kt
@@ -1,0 +1,7 @@
+package br.com.m3Tech.solucoesFromtis.repositories
+
+import br.com.m3Tech.solucoesFromtis.model.ConfGlobal
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ConfGlobalRepository : JpaRepository<ConfGlobal, Integer> {
+}
