+++ src/main/java/br/com/m3Tech/solucoesFromtis/service/IBaseService.java
@@ -5,17 +5,15 @@ import java.util.List;
 import br.com.m3Tech.solucoesFromtis.model.Base;
 
 
-public interface IBaseService {
+ public interface IBaseService {
 	
-	public void salvar(Base base) throws Exception;
+	 void salvar(Base base) throws Exception;
 
-	public List<Base> findAll();
+	 List<Base> findAll();
 
-	public void delete(Base base);
+	 void delete(Base base);
 
-	public Base findById(String id);
-
-	public Base findById(Integer id);
+	 Base findById(Integer id);
 	
 
 }
