+++ src/main/java/br/com/m3Tech/solucoesFromtis/service/impl/BaseServiceImpl.java
@@ -3,70 +3,38 @@ package br.com.m3Tech.solucoesFromtis.service.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import br.com.m3Tech.solucoesFromtis.repositories.BaseRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import br.com.m3Tech.solucoesFromtis.model.Base;
 import br.com.m3Tech.solucoesFromtis.service.IBaseService;
 
 @Service
+@RequiredArgsConstructor
 public class BaseServiceImpl implements IBaseService {
 
-	@Override
-	public void salvar(Base base) throws Exception {
+    private final BaseRepository baseRepository;
 
-//		base.save(); //TODO
-		
-	}
+    @Override
+    public void salvar(Base base) throws Exception {
+        this.baseRepository.save(base);
+    }
 
-	@Override
-	public List<Base> findAll() {
-/*TODO
-   		try {
-			return new Base().findAll();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}*/
-		
-		return new ArrayList<Base>();
-	}
+    @Override
+    public List<Base> findAll() {
+        return this.baseRepository.findAll();
+    }
 
-	@Override
-	public void delete(Base base) {
-	/* TODO
-		try {
-			base.delete();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}*/
-		
-	}
+    @Override
+    public void delete(Base base) {
+        this.baseRepository.delete(base);
+    }
 
-	@Override
-	public Base findById(String id) {
-		//TODO
-	/*	try {
-			return new Base().findById(Integer.valueOf(id));
-		} catch (Exception e) {
-			e.printStackTrace();
-		}*/
-		
-		
-		return null;
-	}
-	
-	@Override
-	public Base findById(Integer id) {
-		//TODO
-	/*	try {
-			return new Base().findById(id);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}*/
-		
-		
-		return null;
-	}
+    @Override
+    public Base findById(Integer id) {
+        return this.baseRepository.findById(id).orElse(null);
+    }
 
-	
 
 }
