+++ src/main/resources/application.properties
@@ -12,4 +12,3 @@ server.servlet.context-path=/gerador
 spring.h2.console.enabled=true
 
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.naming.implicit-strategy=
